??? direct edits to pathfinder

+++ src/main/include/pathfinder/followers/encoder.h
@@ -2,7 +2,7 @@
 #define PATHFINDER_FOL_ENCODER_H_DEF
 
 typedef struct {
-    int initial_position, ticks_per_revolution;
+    double initial_position, ticks_per_revolution;
     double wheel_circumference;
     double kp, ki, kd, kv, ka;
 } EncoderConfig;
@@ -12,8 +12,8 @@ typedef struct {
     int segment, finished;
 } EncoderFollower;
 
-double pathfinder_follow_encoder(EncoderConfig c, EncoderFollower *follower, Segment *trajectory, int trajectory_length, int encoder_tick);
+double pathfinder_follow_encoder(EncoderConfig c, EncoderFollower *follower, Segment *trajectory, int trajectory_length, double encoder_tick);
 
-double pathfinder_follow_encoder2(EncoderConfig c, EncoderFollower *follower, Segment segment, int trajectory_length, int encoder_tick);
+double pathfinder_follow_encoder2(EncoderConfig c, EncoderFollower *follower, Segment segment, int trajectory_length, double encoder_tick);
 
 #endif
